+++ variables.tf
@@ -709,60 +709,37 @@ variable "key_data" {
 
 variable "enable_cluster_log_monitor_diagnostic" {
   type        = bool
-  description = "(optional) Whether to enable the log monitor diagnostic for flexible postgresql server or not?"
+  description = "(optional) Whether to enable the log monitor diagnostic for Azure Kubernetes cluster or not?"
   default     = false
 }
 
 variable "enable_cluster_all_category_group_log" {
   type        = bool
-  description = "(optional) Whether to enable all category group log for flexible postgresql cluster or not?"
+  description = "(optional) Whether to enable all category group log for Azure Kubernetes cluster or not?"
   default     = false
 }
 
 variable "enable_cluster_audit_category_group_log" {
   type        = string
-  description = "(optional) Whether to enable only audit category group log for flexible postgresql cluster or not?"
+  description = "(optional) Whether to enable only audit category group log for Azure Kubernetes cluster or not?"
   default     = true
 }
 
 variable "enable_cluster_custom_categories_log" {
   type        = string
-  description = "(optional) Whether to enable custom categories log for flexible postgresql cluster or not?"
+  description = "(optional) Whether to enable custom categories log for Azure Kubernetes cluster or not?"
   default     = false
-}
-
-variable "enable_cluster_log_retention_policy" {
-  type        = bool
-  description = "(optional) Whether to enable log retention policy for flexible postgresql cluster or not?"
-  default     = true
-}
-variable "cluster_log_retention_policy_days" {
-  type        = string
-  description = "(optional) The number of days for which this Retention Policy should apply for logs for flexible postgresql cluster."
-  default     = 30
 }
 
 variable "enable_cluster_all_metrics" {
   type        = bool
-  description = "(optional) Whether to enable all metrics for diagnostics for flexible postgresql cluster or not?"
+  description = "(optional) Whether to enable all metrics for diagnostics for Azure Kubernetes cluster or not?"
   default     = false
-}
-
-variable "enable_cluster_metric_retention_policy" {
-  type        = string
-  description = "(optional) Whether to enable metric retention policy for flexible postgresql cluster or not?"
-  default     = true
-}
-
-variable "cluster_metric_retention_policy_days" {
-  type        = string
-  description = "(optional) The number of days for which this Retention Policy should apply for metrics for flexible postgresql cluster."
-  default     = 30
 }
 
 variable "cluster_custom_log_categories" {
   type        = list(string)
-  description = "(optional) A list of strings with supported category groups for postgresql cluster log monitoring diagnostics"
+  description = "(optional) A list of strings with supported category groups for Azure Kubernetes cluster log monitoring diagnostics"
   default     = []
 }
 variable "monitor_diagnostic_storage_account_id" {